Revert "Merge pull request #347 from homebase-id/group-chat-fixes"
This reverts commit f3611948f635b3020f88fbffe50980dd3bc1d92d, reversing
changes made to 01cb6e1f49c7372c2b1d12b9cdc4d0af19e88385.
