--- homebase-api/src/commonMain/kotlin/id/homebase/api/client/connections/IntroductionSender.kt
@@ -1,5 +1,0 @@
-package id.homebase.api.client.connections
-
-interface IntroductionSender {
-    suspend fun sendIntroductions(group: IntroductionGroup): IntroductionResult
-}
