+++ homebase-chat/src/commonMain/kotlin/id/homebase/chat/services/ChatMessageSenderService.kt
@@ -47,7 +47,7 @@ class ChatMessageSenderService(
     private val fileOperationsProvider: FileOperationsProvider,
     private val driveFileProvider: DriveFileProvider,
     private val shareSuggestionDonor: ShareSuggestionDonor = ShareSuggestionDonor(),
-) : id.homebase.chat.services.convo.StatusMessageSender {
+) {
     companion object {
         private const val TAG = "ChatMessageSenderService"
     }
@@ -120,13 +120,13 @@ class ChatMessageSenderService(
     }
 
 
-    override suspend fun sendStatusMessage(
+    suspend fun sendStatusMessage(
         messageUniqueId: Uuid,
         conversationId: Uuid,
         statusMessage: StatusMessageData,
-        previousMessageUniqueId: Uuid?,
-        payloadBundle: PayloadBundle?,
-        additionalRecipients: List<OdinId>
+        previousMessageUniqueId: Uuid? = null,
+        payloadBundle: PayloadBundle? = null,
+        additionalRecipients: List<OdinId> = emptyList()
     ): SendMessageResult = sendMessageInternal(
         messageUniqueId = messageUniqueId,
         conversationId = conversationId,
