--- homebase-chat/src/jvmTest/kotlin/id/homebase/chat/services/convo/ConversationServiceRecoveryTest.kt
@@ -1,232 +1,0 @@
-package id.homebase.chat.services.convo
-
-import id.homebase.api.client.drives.files.ArchivalStatus
-import id.homebase.api.common.OdinId
-import id.homebase.chat.services.ChatProtocol
-import id.homebase.chat.services.XorIdUtil
-import kotlin.test.Test
-import kotlin.test.assertEquals
-import kotlin.test.assertFailsWith
-import kotlin.test.assertNotNull
-import kotlin.test.assertNull
-import kotlin.test.assertTrue
-import kotlin.uuid.Uuid
-import kotlinx.coroutines.test.runTest
-
-/**
- * Tier-3 coverage for [ConversationService.recoverConversation] and the admin
- * accessors ([ConversationService.getAdmins],
- * [ConversationService.getConversationAdminHomebaseFile]).
- */
-class ConversationServiceRecoveryTest {
-
-    // ---------- recoverConversation ----------
-
-    @Test
-    fun recoverConversation_noteToSelf_delegatesToEnsureNoteToSelfExists() = runTest {
-        ConversationServiceTestFixture().use { fixture ->
-            val service = fixture.build(scope = this)
-            val rowsBefore = fixture.outboxRowCount()
-
-            service.recoverConversation(
-                conversationId = ChatProtocol.ConversationWithYourselfId,
-                originalAuthor = OdinId(fixture.testDomain),
-            )
-
-            // No existing note-to-self file ⇒ ensureNoteToSelfExists creates + pins.
-            assertTrue(
-                fixture.outboxRowCount() > rowsBefore,
-                "note-to-self recovery should enqueue a creation/pin",
-            )
-        }
-    }
-
-    @Test
-    fun recoverConversation_oneOnOne_existingActive_isNoOp() = runTest {
-        ConversationServiceTestFixture().use { fixture ->
-            val service = fixture.build(scope = this)
-            val alice = "alice.test"
-            val xorId = XorIdUtil.getNewXorId(fixture.testDomain, alice)
-            fixture.seedOneOnOne(other = alice, conversationId = xorId)
-            val rowsBefore = fixture.outboxRowCount()
-
-            service.recoverConversation(
-                conversationId = xorId,
-                originalAuthor = OdinId(alice),
-            )
-
-            assertEquals(
-                rowsBefore,
-                fixture.outboxRowCount(),
-                "healthy active 1:1 needs no recovery ⇒ no new outbox rows",
-            )
-        }
-    }
-
-    @Test
-    fun recoverConversation_oneOnOne_existingDeleted_revivesViaUpdate() = runTest {
-        ConversationServiceTestFixture().use { fixture ->
-            val service = fixture.build(scope = this)
-            val alice = "alice.test"
-            val xorId = XorIdUtil.getNewXorId(fixture.testDomain, alice)
-            fixture.seedOneOnOne(
-                other = alice,
-                conversationId = xorId,
-                archivalStatus = ArchivalStatus.Removed.value,
-            )
-            val rowsBefore = fixture.outboxRowCount()
-
-            service.recoverConversation(
-                conversationId = xorId,
-                originalAuthor = OdinId(alice),
-            )
-
-            assertTrue(
-                fixture.outboxRowCount() > rowsBefore,
-                "deleted 1:1 should be revived via an UpdateFileByUniqueId enqueue",
-            )
-        }
-    }
-
-    @Test
-    fun recoverConversation_oneOnOne_noFile_createsFresh() = runTest {
-        ConversationServiceTestFixture().use { fixture ->
-            val service = fixture.build(scope = this)
-            val alice = "alice.test"
-            val xorId = XorIdUtil.getNewXorId(fixture.testDomain, alice)
-            val rowsBefore = fixture.outboxRowCount()
-
-            service.recoverConversation(
-                conversationId = xorId,
-                originalAuthor = OdinId(alice),
-            )
-
-            assertTrue(
-                fixture.outboxRowCount() > rowsBefore,
-                "no local file ⇒ writeConversationFile must enqueue a new-file upload",
-            )
-        }
-    }
-
-    @Test
-    fun recoverConversation_group_noFile_createsWithCallerAndOriginalAuthor() = runTest {
-        ConversationServiceTestFixture().use { fixture ->
-            val service = fixture.build(scope = this)
-            val alice = "alice.test"
-            // Random uuid that does NOT match XorIdUtil(domain, alice) ⇒ group branch.
-            val groupId = Uuid.random()
-            val rowsBefore = fixture.outboxRowCount()
-
-            service.recoverConversation(
-                conversationId = groupId,
-                originalAuthor = OdinId(alice),
-            )
-
-            assertTrue(
-                fixture.outboxRowCount() > rowsBefore,
-                "group recovery with no file should enqueue a new group conversation",
-            )
-        }
-    }
-
-    // ---------- getAdmins ----------
-
-    @Test
-    fun getAdmins_returnsAdminsFromAdminFile_whenPresent() = runTest {
-        ConversationServiceTestFixture().use { fixture ->
-            val service = fixture.build(scope = this)
-            val alice = "alice.test"
-            val bob = "bob.test"
-            val groupId = fixture.seedGroup(
-                others = listOf(alice, bob),
-                adminDomains = listOf(alice, bob),
-            )
-
-            val admins = service.getAdmins(groupId)
-
-            assertEquals(setOf(OdinId(alice), OdinId(bob)), admins)
-        }
-    }
-
-    @Test
-    fun getAdmins_fallsBackToOriginalAuthor_whenNoAdminFile() = runTest {
-        ConversationServiceTestFixture().use { fixture ->
-            val service = fixture.build(scope = this)
-            val alice = "alice.test"
-            // seedGroup with seedAdminFile = false skips the dedicated admin file.
-            val groupId = fixture.seedGroup(
-                others = listOf(alice),
-                adminDomains = listOf(fixture.testDomain),
-                seedAdminFile = false,
-            )
-
-            val admins = service.getAdmins(groupId)
-
-            // originalAuthor on the seeded conversation is testDomain.
-            assertEquals(setOf(OdinId(fixture.testDomain)), admins)
-        }
-    }
-
-    // ---------- getConversationAdminHomebaseFile ----------
-
-    @Test
-    fun getConversationAdminHomebaseFile_returnsNull_whenNoAdminFile() = runTest {
-        ConversationServiceTestFixture().use { fixture ->
-            val service = fixture.build(scope = this)
-            val groupId = fixture.seedGroup(
-                others = listOf("alice.test"),
-                adminDomains = listOf(fixture.testDomain),
-                seedAdminFile = false,
-            )
-
-            val adminFile = service.getConversationAdminHomebaseFile(groupId)
-
-            assertNull(adminFile)
-        }
-    }
-
-    @Test
-    fun getConversationAdminHomebaseFile_returnsFile_whenPresent() = runTest {
-        ConversationServiceTestFixture().use { fixture ->
-            val service = fixture.build(scope = this)
-            val groupId = fixture.seedGroup(
-                others = listOf("alice.test"),
-                adminDomains = listOf(fixture.testDomain),
-                // seedAdminFile defaults to true
-            )
-
-            val adminFile = service.getConversationAdminHomebaseFile(groupId)
-
-            assertNotNull(adminFile)
-            assertEquals(
-                ChatProtocol.ConversationAdminFileType,
-                adminFile.fileMetadata.appData.fileType,
-            )
-        }
-    }
-
-    // ---------- small accessors ----------
-
-    @Test
-    fun getConversation_returnsNull_forUnknownId() = runTest {
-        ConversationServiceTestFixture().use { fixture ->
-            val service = fixture.build(scope = this)
-
-            val result = service.getConversation(Uuid.random())
-
-            assertNull(result)
-        }
-    }
-
-    @Test
-    fun requireConversation_throws_forUnknownId() = runTest {
-        ConversationServiceTestFixture().use { fixture ->
-            val service = fixture.build(scope = this)
-
-            val ex = assertFailsWith<IllegalStateException> {
-                service.requireConversation(Uuid.random())
-            }
-            assertTrue(ex.message!!.contains("No conversation"))
-        }
-    }
-}
