--- homebase-chat/src/jvmTest/kotlin/id/homebase/chat/services/convo/ConversationServiceTestFixture.kt
@@ -1,460 +1,0 @@
-package id.homebase.chat.services.convo
-
-import app.cash.sqldelight.driver.jdbc.sqlite.JdbcSqliteDriver
-import id.homebase.api.client.auth.ApiCredentials
-import id.homebase.api.client.auth.CredentialsManager
-import id.homebase.api.client.connections.IntroductionGroup
-import id.homebase.api.client.connections.IntroductionResult
-import id.homebase.api.client.connections.IntroductionSender
-import id.homebase.api.client.drives.HomebaseFile
-import id.homebase.api.client.eventbus.EventBus
-import id.homebase.api.common.OdinId
-import id.homebase.api.common.SecureByteArray
-import id.homebase.api.serialization.OdinSystemSerializer
-import id.homebase.api.sync.database.DatabaseManager
-import id.homebase.api.sync.database.MainIndexMetaHelpers
-import id.homebase.api.sync.database.OdinDatabase
-import id.homebase.api.sync.database.Outbox
-import id.homebase.api.sync.database.OutboxSync
-import id.homebase.api.sync.database.OutboxUploader
-import id.homebase.chat.services.ChatProtocol
-import id.homebase.chat.services.PayloadBundle
-import id.homebase.chat.services.PayloadBundleEncryptor
-import id.homebase.chat.services.SendMessageResult
-import id.homebase.chat.services.StatusMessageData
-import id.homebase.chat.services.outbox.OptimisticWriter
-import kotlin.time.Clock
-import kotlin.uuid.Uuid
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.test.TestScope
-
-/**
- * Shared setup for ConversationService tests.
- *
- *   - Real in-memory SQLite DB (same pattern as [AdminQueryTest]).
- *   - Real [OutboxSync] with `setOnline(false)` so enqueued rows land in the DB for
- *     inspection but are never actually uploaded. The stub [OutboxUploader] throws
- *     to catch accidental online paths in tests.
- *   - Fake collaborators ([FakeStatusMessageSender], [FakeIntroductionSender],
- *     [FakeConversationLoader], [FakePayloadBundleEncryptor]) record calls for
- *     assertion — no network, no crypto, no coroutines they can leak.
- *
- * Use via `ConversationServiceTestFixture().use { fixture -> ... }` so the
- * in-memory DB is closed at the end of each test.
- */
-class ConversationServiceTestFixture : AutoCloseable {
-
-    // Fixed identityId — matches the hardcoded value in ApiCredentials.getIdentityId()
-    val testIdentityId: Uuid = Uuid.parse("7b1be23b-48bb-4304-bc7b-db5910c09a92")
-
-    // Chat drive alias from SystemDriveConstants.chatDrive
-    val chatDriveId: Uuid = Uuid.parse("9ff813af-f2d6-1e2f-9b9d-b189e72d1a11")
-
-    val testDomain: String = "owner.test"
-
-    lateinit var dbm: DatabaseManager
-        private set
-    lateinit var credentialsManager: CredentialsManager
-        private set
-    lateinit var eventBus: EventBus
-        private set
-    lateinit var outboxSync: OutboxSync
-        private set
-    lateinit var introductionSender: FakeIntroductionSender
-        private set
-    lateinit var statusMessageSender: FakeStatusMessageSender
-        private set
-    lateinit var conversationLoader: FakeConversationLoader
-        private set
-    lateinit var payloadEncryptor: FakePayloadBundleEncryptor
-        private set
-
-    suspend fun build(scope: CoroutineScope = TestScope()): ConversationService {
-        dbm = createInMemoryDbm()
-        credentialsManager = createCredentialsManager(testDomain)
-        eventBus = EventBus()
-        outboxSync = OutboxSync(
-            databaseManager = dbm,
-            uploader = ThrowingOutboxUploader,
-            eventBus = eventBus,
-            scope = scope
-        ).also { it.setOnline(false) } // never actually send
-
-        introductionSender = FakeIntroductionSender()
-        statusMessageSender = FakeStatusMessageSender()
-        conversationLoader = FakeConversationLoader()
-        payloadEncryptor = FakePayloadBundleEncryptor()
-
-        val optimisticWriter = OptimisticWriter(
-            credentialsManager = credentialsManager,
-            dbm = dbm,
-            eventBus = eventBus
-        )
-
-        return ConversationService(
-            credentialsManager = credentialsManager,
-            payloadBundleEncryptionService = payloadEncryptor,
-            dbm = dbm,
-            introductionProvider = introductionSender,
-            scope = scope,
-            outboxSync = outboxSync,
-            chatMessageSenderService = statusMessageSender,
-            optimisticWriter = optimisticWriter,
-            conversationStream = conversationLoader,
-        )
-    }
-
-    // ---------- DB seed helpers ----------
-
-    /**
-     * Seed a group conversation where [testDomain] is the admin. Participants include
-     * the caller ([testDomain]) plus every entry in [others].
-     *
-     * @param localTags tags written to localAppData — drive Left/Archived/Pinned state.
-     * @param archivalStatus 0=Active, 2=Removed. See [ArchivalStatus].
-     */
-    suspend fun seedGroup(
-        conversationId: Uuid = Uuid.random(),
-        others: List<String>,
-        adminDomains: List<String> = listOf(testDomain),
-        title: String = "",
-        localTags: List<Uuid> = emptyList(),
-        archivalStatus: Int = 0,
-        seedAdminFile: Boolean = true,
-    ): Uuid {
-        val participants = listOf(testDomain) + others
-        val adminDataJson =
-            """{"admins":[${adminDomains.joinToString(",") { "\"$it\"" }}]}"""
-        insertConversationFile(
-            fileId = Uuid.random(),
-            uniqueId = conversationId,
-            participants = participants,
-            originalAuthor = testDomain,
-            isGroup = true,
-            adminDataJson = adminDataJson,
-            title = title,
-            localTags = localTags,
-            archivalStatus = archivalStatus,
-        )
-
-        if (seedAdminFile) {
-            // Separate admin file (takes priority over adminData when read by
-            // ConversationAdminInfo.queryFromDb — see AdminQueryTest for details).
-            insertAdminFile(
-                fileId = Uuid.random(),
-                uniqueId = ChatProtocol.getAdminFileUniqueId(conversationId),
-                groupId = conversationId,
-                admins = adminDomains,
-                originalAuthor = testDomain
-            )
-        }
-        return conversationId
-    }
-
-    /** Seed a 1:1 conversation between [testDomain] and [other]. */
-    suspend fun seedOneOnOne(
-        other: String,
-        conversationId: Uuid = Uuid.random(),
-        localTags: List<Uuid> = emptyList(),
-        archivalStatus: Int = 0,
-    ): Uuid {
-        insertConversationFile(
-            fileId = Uuid.random(),
-            uniqueId = conversationId,
-            participants = listOf(testDomain, other),
-            originalAuthor = testDomain,
-            isGroup = false,
-            title = "",
-            localTags = localTags,
-            archivalStatus = archivalStatus,
-        )
-        return conversationId
-    }
-
-    /** A legacy (pre-tag) group — >2 participants but no ConversationGroupTag. */
-    suspend fun seedLegacyGroup(
-        others: List<String>,
-        conversationId: Uuid = Uuid.random(),
-    ): Uuid {
-        insertConversationFile(
-            fileId = Uuid.random(),
-            uniqueId = conversationId,
-            participants = listOf(testDomain) + others,
-            originalAuthor = testDomain,
-            isGroup = false, // no ConversationGroupTag → legacy
-            title = "",
-        )
-        return conversationId
-    }
-
-    fun outboxRowCount(): Long = dbm.outbox.count()
-
-    /**
-     * Drain the outbox (checkout everything eligible at real `now`) so callers can
-     * inspect what was enqueued. NOTE: this mutates `checkOutStamp`. Don't call it
-     * if you plan to run further service operations that might re-enqueue the same
-     * uniqueId, and don't rely on [outboxRowCount] afterwards — the rows are still
-     * there, just checked out.
-     */
-    suspend fun drainOutbox(): List<Outbox> {
-        val drained = mutableListOf<Outbox>()
-        while (true) {
-            val row = dbm.outbox.checkout() ?: break
-            drained += row
-        }
-        return drained
-    }
-
-    /** Load the current conversation HomebaseFile from the in-memory DB. */
-    suspend fun getConversationFile(conversationId: Uuid) =
-        dbm.driveMainIndex.selectHomebaseFileByUnique(
-            testIdentityId, chatDriveId, conversationId
-        )
-
-    override fun close() {
-        if (::dbm.isInitialized) dbm.close()
-    }
-
-    // ---------- internals ----------
-
-    private fun createInMemoryDbm(): DatabaseManager = DatabaseManager {
-        val driver = JdbcSqliteDriver(JdbcSqliteDriver.IN_MEMORY)
-        OdinDatabase.Schema.create(driver)
-        driver
-    }
-
-    private suspend fun createCredentialsManager(domain: String): CredentialsManager {
-        val cm = CredentialsManager()
-        cm.setActiveCredentials(
-            ApiCredentials.create(
-                domain = OdinId(domain),
-                clientAccessToken = "test-token",
-                sharedSecret = SecureByteArray(ByteArray(16))
-            )
-        )
-        return cm
-    }
-
-    private suspend fun insertConversationFile(
-        fileId: Uuid,
-        uniqueId: Uuid,
-        participants: List<String>,
-        originalAuthor: String,
-        isGroup: Boolean,
-        adminDataJson: String? = null,
-        title: String = "",
-        archivalStatus: Int = 0,
-        localTags: List<Uuid> = emptyList(),
-    ) {
-        val now = Clock.System.now().epochSeconds
-        val recipientsJson = participants.joinToString(",") { "\"$it\"" }
-        val tagsField = if (isGroup) "[\"${ChatProtocol.ConversationGroupTag}\"]" else "null"
-        val adminDataField = if (adminDataJson != null) """, "adminData": $adminDataJson""" else ""
-        val contentJson =
-            """{"title":"$title","version":1,"recipients":[$recipientsJson]$adminDataField}"""
-        val escaped = contentJson.replace("\"", "\\\"")
-        val localAppDataField = if (localTags.isEmpty()) {
-            "null"
-        } else {
-            val tagList = localTags.joinToString(",") { "\"$it\"" }
-            """{"tags":[$tagList],"versionTag":"${Uuid.random()}","iv":null,"content":null,"readTime":null,"localReactions":null}"""
-        }
-
-        insertFile(
-            """{
-              "fileId": "$fileId",
-              "driveId": "$chatDriveId",
-              "fileState": "active",
-              "fileSystemType": "standard",
-              "serverFileIsEncrypted": "false",
-              "keyHeader": {
-                "iv": [0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0],
-                "aesKey": {"bytes": [0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0]}
-              },
-              "fileMetadata": {
-                "globalTransitId": "${Uuid.random()}",
-                "created": ${now}000,
-                "updated": ${now}000,
-                "transitCreated": ${now}000,
-                "transitUpdated": 0,
-                "isEncrypted": false,
-                "senderOdinId": "$originalAuthor",
-                "originalAuthor": "$originalAuthor",
-                "appData": {
-                  "uniqueId": "$uniqueId",
-                  "tags": $tagsField,
-                  "fileType": ${ChatProtocol.ConversationFileType},
-                  "dataType": 0,
-                  "groupId": null,
-                  "userDate": ${now}000,
-                  "content": "$escaped",
-                  "previewThumbnail": null,
-                  "archivalStatus": $archivalStatus
-                },
-                "localAppData": $localAppDataField,
-                "referencedFile": null,
-                "reactionPreview": null,
-                "versionTag": "${Uuid.random()}",
-                "payloads": [],
-                "dataSource": null
-              },
-              "serverMetadata": {
-                "accessControlList": {
-                  "requiredSecurityGroup": "owner",
-                  "circleIdList": null,
-                  "odinIdList": null
-                },
-                "doNotIndex": false,
-                "allowDistribution": true,
-                "fileSystemType": "standard",
-                "fileByteCount": 100,
-                "originalRecipientCount": ${participants.size - 1},
-                "transferHistory": null
-              },
-              "priority": 300,
-              "fileByteCount": 100
-            }"""
-        )
-    }
-
-    private suspend fun insertAdminFile(
-        fileId: Uuid,
-        uniqueId: Uuid,
-        groupId: Uuid,
-        admins: List<String>,
-        originalAuthor: String,
-    ) {
-        val now = Clock.System.now().epochSeconds
-        val adminsJson = admins.joinToString(",") { "\"$it\"" }
-        val contentJson = """{"admins":[$adminsJson]}"""
-        val escaped = contentJson.replace("\"", "\\\"")
-
-        insertFile(
-            """{
-              "fileId": "$fileId",
-              "driveId": "$chatDriveId",
-              "fileState": "active",
-              "fileSystemType": "standard",
-              "serverFileIsEncrypted": "false",
-              "keyHeader": {
-                "iv": [0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0],
-                "aesKey": {"bytes": [0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0]}
-              },
-              "fileMetadata": {
-                "globalTransitId": "${Uuid.random()}",
-                "created": ${now}000,
-                "updated": ${now}000,
-                "transitCreated": 0,
-                "transitUpdated": 0,
-                "isEncrypted": false,
-                "senderOdinId": "$originalAuthor",
-                "originalAuthor": "$originalAuthor",
-                "appData": {
-                  "uniqueId": "$uniqueId",
-                  "tags": null,
-                  "fileType": ${ChatProtocol.ConversationAdminFileType},
-                  "dataType": 0,
-                  "groupId": "$groupId",
-                  "userDate": ${now}000,
-                  "content": "$escaped",
-                  "previewThumbnail": null,
-                  "archivalStatus": 0
-                },
-                "localAppData": null,
-                "referencedFile": null,
-                "reactionPreview": null,
-                "versionTag": "${Uuid.random()}",
-                "payloads": [],
-                "dataSource": null
-              },
-              "serverMetadata": {
-                "accessControlList": {
-                  "requiredSecurityGroup": "owner",
-                  "circleIdList": null,
-                  "odinIdList": null
-                },
-                "doNotIndex": false,
-                "allowDistribution": true,
-                "fileSystemType": "standard",
-                "fileByteCount": 50,
-                "originalRecipientCount": 0,
-                "transferHistory": null
-              },
-              "priority": 300,
-              "fileByteCount": 50
-            }"""
-        )
-    }
-
-    private suspend fun insertFile(jsonHeader: String) {
-        val header = OdinSystemSerializer.deserialize<HomebaseFile>(jsonHeader)
-        val processor = MainIndexMetaHelpers.HomebaseFileProcessor(dbm)
-        val record = processor.convertFileHeaderToDriveMainIndexRecord(
-            testIdentityId, chatDriveId, header
-        )
-        MainIndexMetaHelpers.upsertDriveMainIndex(dbm, record)
-    }
-}
-
-// ---------- Test doubles ----------
-
-class FakeIntroductionSender : IntroductionSender {
-    val calls = mutableListOf<IntroductionGroup>()
-    override suspend fun sendIntroductions(group: IntroductionGroup): IntroductionResult {
-        calls += group
-        return IntroductionResult()
-    }
-}
-
-class FakeStatusMessageSender : StatusMessageSender {
-    data class Call(
-        val conversationId: Uuid,
-        val statusMessage: StatusMessageData,
-        val additionalRecipients: List<OdinId>,
-    )
-    val calls = mutableListOf<Call>()
-    override suspend fun sendStatusMessage(
-        messageUniqueId: Uuid,
-        conversationId: Uuid,
-        statusMessage: StatusMessageData,
-        previousMessageUniqueId: Uuid?,
-        payloadBundle: PayloadBundle?,
-        additionalRecipients: List<OdinId>,
-    ): SendMessageResult {
-        calls += Call(conversationId, statusMessage, additionalRecipients)
-        return SendMessageResult(messageUniqueId)
-    }
-}
-
-class FakeConversationLoader : ConversationLoader {
-    val loaded = mutableListOf<Uuid>()
-    override suspend fun loadConversation(conversationId: Uuid) {
-        loaded += conversationId
-    }
-}
-
-class FakePayloadBundleEncryptor : PayloadBundleEncryptor {
-    val calls = mutableListOf<Uuid>()
-    override suspend fun encryptBundle(
-        uniqueId: Uuid,
-        bundle: PayloadBundle?,
-        aesKey: SecureByteArray,
-        scope: CoroutineScope,
-    ): PayloadBundle {
-        calls += uniqueId
-        // Empty bundle regardless of input. Tests that need to assert payload contents
-        // should use a more specific fake or the real service.
-        return PayloadBundle(
-            payloads = emptyList(),
-            thumbnails = emptyList(),
-            previewThumbs = emptyList(),
-        )
-    }
-}
-
-/** Fails if the outbox ever tries to send — ensures tests stay offline. */
-private object ThrowingOutboxUploader : OutboxUploader {
-    override suspend fun upload(outboxRecord: Outbox, eventBus: EventBus) {
-        error("OutboxUploader.upload should never be called in tests (setOnline(false))")
-    }
-}
